+++ virotour/lib/src/text_search/view_object.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
-import 'package:http/http.dart' as http;
+
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 
 class ViewObject extends StatelessWidget {
   String url;
